nixos/acme: Service should not fail when no renewal is necessary
The acme service script re-uses the exit code of the simp_le program
which is 0 on successful change, 1 if no renewal is necessary, and 2 on
error. This causes the service to end up in failed status when no
renewal is necessary, which is wrong.

With this patch the exit code from the acme script is 0 when simp_le
returns 1 and otherwise the same as returned by simp_le.

The postStop script still uses the unchanged exit code from simp_le to
determine wether to run the postRun hook or not.